Move `assert_data_tables_equal`.
