+++ tests_metricflow/table_snapshot/test_table_snapshots.py
@@ -10,8 +10,8 @@ from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfi
 
 from metricflow.data_table.mf_table import MetricFlowDataTable
 from metricflow.protocols.sql_client import SqlEngine
-from tests_metricflow.compare_df import assert_data_tables_equal
 from tests_metricflow.fixtures.sql_clients.ddl_sql_client import SqlClientWithDDLMethods
+from tests_metricflow.sql.compare_data_table import assert_data_tables_equal
 from tests_metricflow.table_snapshot.table_snapshots import (
     SqlTableColumnDefinition,
     SqlTableColumnType,
